+++ site/dat/docs/Changelog.md
@@ -10,6 +10,7 @@
  - fix cumulative part of BlazeMeter reporting
  - remove MirrorsManager from Gatling
  - add final stats reporting for functional mode
+ - add `content-encoding` request- and scenario-level option to JMeter
 
 ## 1.7.2 <sup>13 oct 2016</sup>
  - fix keep-alive processing in Gatling